+++ packages/core/src/components/ui/LocalizationButton/LocalizationButton.tsx
@@ -6,6 +6,7 @@ import { Icon, Button as UIButton } from '@faststore/ui'
 import storeConfig from 'discovery.config'
 import LocalizationSelector from 'src/components/localization/LocalizationSelector'
 import { useBindingSelector } from 'src/sdk/localization'
+import { useLocalizedProduct } from 'src/sdk/localization/LocalizedProductContext'
 import { useSession } from 'src/sdk/session'
 
 export interface LocalizationButtonErrorMessages {
@@ -41,6 +42,9 @@ const LocalizationButton = ({
   const [isSelectorOpen, setIsSelectorOpen] = useState(false)
   const buttonRef = useRef<HTMLButtonElement>(null)
 
+  const otherLocales = useLocalizedProduct()?.otherLocales ?? undefined
+  const urlSuffix = useLocalizedProduct()?.urlSuffix ?? '/p'
+
   const {
     languages,
     currencies,
@@ -52,7 +56,7 @@ const LocalizationButton = ({
     reset,
     isSaveEnabled,
     error,
-  } = useBindingSelector()
+  } = useBindingSelector(otherLocales, urlSuffix)
 
   const { locale: sessionLocale, currency: sessionCurrency } = useSession()